refactor avro reader and writer
