+++ mr/src/main/java/org/apache/iceberg/mr/mapreduce/IcebergInputFormat.java
@@ -53,7 +53,7 @@ import org.apache.iceberg.catalog.Catalog;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.common.DynConstructors;
 import org.apache.iceberg.data.IdentityPartitionConverters;
-import org.apache.iceberg.data.avro.DataReader;
+import org.apache.iceberg.data.avro.AvroGenericReader;
 import org.apache.iceberg.data.orc.GenericOrcReader;
 import org.apache.iceberg.data.parquet.GenericParquetReaders;
 import org.apache.iceberg.expressions.Evaluator;
@@ -291,7 +291,7 @@ public class IcebergInputFormat<T> extends InputFormat<Void, T> {
         case GENERIC:
           avroReadBuilder.createReaderFunc(
               (expIcebergSchema, expAvroSchema) ->
-                  DataReader.create(expIcebergSchema, expAvroSchema,
+                  AvroGenericReader.create(expIcebergSchema, expAvroSchema,
                       constantsMap(task, IdentityPartitionConverters::convertConstant)));
       }
       return applyResidualFiltering(avroReadBuilder.build(), task.residual(), readSchema);
